api/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-seed==0.3.1
djangorestframework==3.14.0
Faker==19.1.0
install==1.3.5
Pillow==10.0.0
python-barcode==0.15.1
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.4
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.3
